% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{connect}
\alias{connect}
\title{Connect to graph}
\usage{
connect(url = NULL, username = NULL, password = NULL)
}
\arguments{
\item{url=NULL}{<what param does>}

\item{username=NULL}{<what param does>}

\item{password=NULL}{<what param does>}
}
\value{
graph object
}
\description{
<full description>
}
