% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{get_genassoc}
\alias{get_genassoc}
\title{<brief desc>}
\usage{
get_genassoc(variant, id, graph = options()$eveneo4jg)
}
\arguments{
\item{graph}{= options()$eveneo4jg <what param does>}

\item{variant=NULL}{<what param does>}

\item{id=NULL}{<what param does>}
}
\value{
dataframe
}
\description{
<full description>
}
