% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{get_mr}
\alias{get_mr}
\title{Get MR estimates}
\usage{
get_mr(id1, id2, graph = options()$eveneo4jg)
}
\arguments{
\item{id1}{<what param does>}

\item{id2}{<what param does>}

\item{graph}{= options()$eveneo4jg <what param does>}
}
\value{
list
}
\description{
Get MR estimates
}
