% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{get_mrmoe}
\alias{get_mrmoe}
\title{Get MR-MOE estimates}
\usage{
get_mrmoe(id1, id2, graph = options()$eveneo4jg)
}
\arguments{
\item{id1}{<what param does>}

\item{id2}{<what param does>}

\item{graph}{= options()$eveneo4jg <what param does>}
}
\value{
data frame
}
\description{
Get MR-MOE estimates
}
