% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{get_traits}
\alias{get_traits}
\title{Get traits}
\usage{
get_traits(id = NULL, graph = options()$eveneo4jg)
}
\arguments{
\item{graph}{<what param does>}

\item{trait=NULL}{<what param does>}
}
\value{
dataframe
}
\description{
<full description>
}
