% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.r
\name{mrever_to_mr}
\alias{mrever_to_mr}
\title{Create TwoSampleMR dat object}
\usage{
mrever_to_mr(id1, id2)
}
\arguments{
\item{id1}{<what param does>}

\item{id2}{<what param does>}
}
\value{

}
\description{
<full description>
}
