% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{phewas}
\alias{phewas}
\title{Perform phewas}
\usage{
phewas(id, direction = "exposure", graph = options()$eveneo4jg)
}
\arguments{
\item{id}{<what param does>}

\item{graph}{= options()$eveneo4jg <what param does>}

\item{direction="exposure"}{<what param does>}
}
\value{
data frame
}
\description{
<full description>
}
