% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{write_out}
\alias{write_out}
\title{Write as csv for neo4j}
\usage{
write_out(x, basename, header = FALSE)
}
\arguments{
\item{x}{Data frame}

\item{basename}{Filename to write to. Should be gzip}

\item{header}{Whether to include header. If so, it is written as a separate file. Default=FALSE}
}
\description{
Write as csv for neo4j
}
